+++ src/containers/DataSubsectionContainer.tsx
@@ -1,24 +1,26 @@
-import React from "react";
 import { connect, ConnectedProps } from "react-redux";
 import { DataSubsection } from "../components/DataSubsection";
-import { IRootState } from "../types";
+import { IRootState } from "../types/redux";
 
 export interface ISubsectionOwnProps {
   datasetProp: string;
   totalProp: string;
 }
 
-const connector = connect(
-  ({ data }: IRootState, { totalProp, datasetProp }: ISubsectionOwnProps) => {
-    const dataset = data[datasetProp] as string;
-    const total = data[totalProp] as number;
+const mapStateToSubsectionProps = (
+  { data }: IRootState,
+  { totalProp, datasetProp }: ISubsectionOwnProps
+) => {
+  const dataset = data[datasetProp] as string;
+  const total = data[totalProp] as number;
 
-    return {
-      dataset,
-      total,
-    };
-  }
-);
+  return {
+    dataset,
+    total,
+  };
+};
+
+const connector = connect(mapStateToSubsectionProps);
 
 export type IDataSubsectionContainer = ConnectedProps<typeof connector>;
