+++ src/types/types.d.ts
@@ -42,18 +42,6 @@ export interface TradesWithTotalSum {
   totalTradesProfitPln: number;
 }
 
-export interface ContextInterface {
-  trades: string;
-  dividends: string;
-  dividendsTotal: number;
-  tradesTotal: number;
-  [k: string]: string | number;
-}
-export type IDataStore = ContextInterface;
-
-export interface IRootState {
-  data: IDataStore;
-}
 export interface ITimeRange {
   endDate: string;
   startDate: string;
